scripts/requirements.txt
python-telegram-bot>=20.7
anthropic>=0.34.0
openai>=1.40.0
python-dotenv>=1.0.0
httpx>=0.27.0
